+++ MacSVG Plugin Projects/macSVG Plugins/PathTextGenerator/PathTextGenerator/PathTextGenerator.h
@@ -20,10 +20,10 @@
 
 - (NSXMLElement *)createPathWithString:(NSString *)pathTextString
         font:(NSFont *)font fontSize:(CGFloat)fontSize
-        originX:(CGFloat)originX originY:(CGFloat)originY;
+        originX:(float)originX originY:(float)originY;
 
 - (NSXMLElement *)createMultiplePathsWithString:(NSString *)pathTextString
         font:(NSFont *)font fontSize:(CGFloat)fontSize
-        originX:(CGFloat)originX originY:(CGFloat)originY;
+        originX:(float)originX originY:(float)originY;
 
 @end
